fix: correct race condition by moving binding into register flow
