+++ src/PaymentGateways/PayPalCommerce/PayPalCommerce.php
@@ -63,22 +63,6 @@ class PayPalCommerce implements PaymentGateway {
 		give( ScriptLoader::class )->boot();
 		give( AjaxRequestHandler::class )->boot();
 
-		give()->singleton(
-			MerchantDetail::class,
-			static function () {
-				return ( new MerchantDetail() )->boot();
-			}
-		);
-
-		give()->singleton( PayPalClient::class );
-
-		give()->singleton(
-			RefreshToken::class,
-			static function() {
-				return ( new RefreshToken() )->boot();
-			}
-		);
-
 		// Boot RefreshToken class immediately if cron job running.
 		if ( wp_doing_cron() ) {
 			give( RefreshToken::class );
